winit-core: partially split event_loop
